[APM] Refactor url handline
 - Remove RelativeLink component
 - Move react component out of url.tsx (renamed to url_helpers.ts)
