+++ x-pack/plugins/apm/public/components/app/ErrorGroupDetails/DetailView/index.tsx
@@ -22,6 +22,7 @@ import {
   ERROR_LOG_STACKTRACE
 } from 'x-pack/plugins/apm/common/constants';
 import { IUrlParams } from 'x-pack/plugins/apm/public/store/urlParams';
+import { KibanaLink } from 'x-pack/plugins/apm/public/utils/url/KibanaLink';
 import { ErrorGroupAPIResponse } from 'x-pack/plugins/apm/server/lib/errors/get_error_group';
 import { APMError } from 'x-pack/plugins/apm/typings/es_schemas/Error';
 import { IStackframe } from 'x-pack/plugins/apm/typings/es_schemas/Stackframe';
@@ -42,8 +43,8 @@ import {
   unit,
   units
 } from '../../../../style/variables';
-import { fromQuery, history, toQuery } from '../../../../utils/url';
-import { KibanaLink, legacyEncodeURIComponent } from '../../../../utils/url';
+import { legacyEncodeURIComponent } from '../../../../utils/url/url_helpers';
+import { fromQuery, history, toQuery } from '../../../../utils/url/url_helpers';
 import { DiscoverErrorButton } from '../../../shared/DiscoverButtons/DiscoverErrorButton';
 import {
   getPropertyTabNames,
