+++ x-pack/plugins/apm/public/components/app/ErrorGroupOverview/List/index.js
@@ -9,12 +9,8 @@ import PropTypes from 'prop-types';
 import { EuiBasicTable, EuiBadge } from '@elastic/eui';
 import numeral from '@elastic/numeral';
 import moment from 'moment';
-import {
-  toQuery,
-  fromQuery,
-  history,
-  RelativeLink
-} from '../../../../utils/url';
+import { toQuery, fromQuery, history } from '../../../../utils/url/url_helpers';
+import { KibanaLink } from '../../../../utils/url/KibanaLink';
 import TooltipOverlay from '../../../shared/TooltipOverlay';
 import styled from 'styled-components';
 import {
@@ -30,7 +26,7 @@ function paginateItems({ items, pageIndex, pageSize }) {
   return items.slice(pageIndex * pageSize, (pageIndex + 1) * pageSize);
 }
 
-const GroupIdLink = styled(RelativeLink)`
+const GroupIdLink = styled(KibanaLink)`
   font-family: ${fontFamilyCode};
 `;
 
@@ -38,7 +34,7 @@ const MessageAndCulpritCell = styled.div`
   ${truncate('100%')};
 `;
 
-const MessageLink = styled(RelativeLink)`
+const MessageLink = styled(KibanaLink)`
   font-family: ${fontFamilyCode};
   font-size: ${fontSizes.large};
   ${truncate('100%')};
@@ -98,7 +94,7 @@ class List extends Component {
         width: px(unit * 6),
         render: groupId => {
           return (
-            <GroupIdLink path={`/${serviceName}/errors/${groupId}`}>
+            <GroupIdLink hash={`/${serviceName}/errors/${groupId}`}>
               {groupId.slice(0, 5) || notAvailableLabel}
             </GroupIdLink>
           );
@@ -118,7 +114,7 @@ class List extends Component {
           return (
             <MessageAndCulpritCell>
               <TooltipOverlay content={message || notAvailableLabel}>
-                <MessageLink path={`/${serviceName}/errors/${item.groupId}`}>
+                <MessageLink hash={`/${serviceName}/errors/${item.groupId}`}>
                   {message || notAvailableLabel}
                 </MessageLink>
               </TooltipOverlay>
