+++ x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/TransactionPropertiesTable.tsx
@@ -12,7 +12,7 @@ import styled from 'styled-components';
 import { Transaction } from '../../../../../typings/es_schemas/Transaction';
 import { IUrlParams } from '../../../../store/urlParams';
 import { px, units } from '../../../../style/variables';
-import { fromQuery, history, toQuery } from '../../../../utils/url';
+import { fromQuery, history, toQuery } from '../../../../utils/url/url_helpers';
 import {
   getPropertyTabNames,
   PropertiesTable,
