+++ x-pack/plugins/apm/public/components/app/TransactionDetails/Transaction/WaterfallContainer/Waterfall/FlyoutTopLevelProperties.tsx
@@ -11,7 +11,7 @@ import {
 } from 'x-pack/plugins/apm/common/constants';
 import { StickyProperties } from 'x-pack/plugins/apm/public/components/shared/StickyProperties';
 import { TransactionLink } from 'x-pack/plugins/apm/public/components/shared/TransactionLink';
-import { KibanaLink } from 'x-pack/plugins/apm/public/utils/url';
+import { KibanaLink } from 'x-pack/plugins/apm/public/utils/url/KibanaLink';
 import { Transaction } from 'x-pack/plugins/apm/typings/es_schemas/Transaction';
 
 interface Props {
