+++ x-pack/plugins/apm/public/components/shared/KueryBar/view.js
@@ -14,7 +14,7 @@ import {
   toQuery,
   legacyEncodeURIComponent,
   KibanaLink
-} from '../../../utils/url';
+} from '../../../utils/url/url_helpers';
 import { Typeahead } from './Typeahead';
 import chrome from 'ui/chrome';
 import {
